支付服务-分库分表&支付逻辑设计

+++ payment-service/payment-server/src/main/java/com/sale/model/PayBill.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.Data;
 
@@ -53,7 +54,7 @@ public class PayBill implements Serializable {
     /**
      * 支付金额
      */
-    private Integer payAmount;
+    private BigDecimal payAmount;
 
     /**
      * 支付种类 详细见枚举PayBillType
